Redesign so PC can read from address bus; not data bus.
Requires removing jump-immediate instructions (JIU,JIZ,JIV)
And instead providing with jump-X instructions (JXU,JXZ,JXV)

Fixup examples to load X before using it to jump.
i.e.
    JIU, IMM somewhere
-->
    LIX, IMM somewhere, JXU

Some examples needed a bit more finessing! In the case that X was already in use to hold a value across the jump.

Also the expected cycles for the examples increased. :(

But at least the HW is simpler!